Add enum settings
Add an enum setting type to allow for dropdown selects of discrete
values

Migrate the debug logging setting to the new enum-based setting

+++ Source/v2/LightsOut2/LightsOut2/Logging/FilteredVerseLogger.cs
@@ -12,7 +12,7 @@ namespace LightsOut2.Logging
 
         protected override void Log(string message, LogLevel logLevel, bool onlyOnce)
         {
-            if ((int)logLevel >= LightsOut2Settings.MinimumLogLevel)
+            if (logLevel >= LightsOut2Settings.MinimumLogLevel)
             {
                 base.Log(message, logLevel, onlyOnce);
             }
